requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.20
redis[hiredis]==5.2.1
httpx==0.28.1
python-dotenv==1.0.1
greenlet==3.1.1
aioimaplib==1.1.0
beautifulsoup4==4.12.3
